+++ CHANGELOG.md
@@ -6,7 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased]
 
-[Unreleased]: https://github.com/LupaHQ/argon-vscode/compare/0.0.25...HEAD
+[Unreleased]: https://github.com/LupaHQ/argon-vscode/compare/0.0.26...HEAD
+
+## [0.0.26] - 2025-04-27
+
+### Changed
+
+- Updated extension to work with renamed Roblox plugin file (now "Lemonade.rbxm")
+- Improved plugin installation and update processes to use the new filename
+
+[0.0.26]: https://github.com/LupaHQ/argon-vscode/compare/0.0.25...0.0.26
 
 ## [0.0.25] - 2024-03-21